Rename python package to graph
