+++ crates/unladen_swallow/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "unladen_swallow"
+name = "pygraph"
 version = "0.2.1"
 authors = [
     "Martin Junghanns <github@s1ck.dev>",
@@ -15,7 +15,7 @@ readme = "../../README.md"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
-name = "unladen_swallow"
+name = "graph"
 crate-type = ["cdylib"]
 
 [features]
