+++ crates/unladen_swallow/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["maturin>=0.13,<0.14"]
 build-backend = "maturin"
 
 [project]
-name = "unladen_swallow"
+name = "graph"
 version = "0.0.1"
 requires-python = ">=3.7"
 classifiers = [
