+++ crates/unladen_swallow/tests/conftest.py
@@ -1,6 +1,6 @@
 from pytest import fixture
 
-from unladen_swallow import DiGraph, Graph
+from graph import DiGraph, Graph
 
 FILE = "../builder/resources/scale_8.graph500"
