+++ crates/unladen_swallow/tests/ds_test.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 
-from unladen_swallow import DiGraph, Graph
+from graph import DiGraph, Graph
 
 
 def test_numpy_graph():
