+++ crates/unladen_swallow/tests/numpy_neighbors_test.py
@@ -1,4 +1,4 @@
-from unladen_swallow import DiGraph, Graph
+from graph import DiGraph, Graph
 
 
 def test_out_neighbors(g: DiGraph):
